biotope website: - updated BioScrollAnimation to be consistent with newer components created with biotope generate 2.0.0
