+++ src/components/BioScrollAnimation/defines.ts
@@ -1,0 +1,36 @@
+/**
+ *  ## DEFINE ALL INTERFACES FOR BioScrollAnimation
+ **/
+
+/**
+ * Props
+ */
+ interface BioScrollAnimationProps {
+	animationDataUrl: string
+	autoplay?: boolean,
+	loop?: boolean,
+	renderer?: string,
+	scrollDuration?: string
+	scrollFactor?: number
+ }
+
+/**
+ * State
+ */
+ interface BioScrollAnimationState {
+
+ }
+
+/**
+ * Methods
+ */
+ interface BioScrollAnimationMethods {
+
+ }
+
+
+export {
+    BioScrollAnimationProps,
+	BioScrollAnimationState,
+	BioScrollAnimationMethods
+}
